MultispectralForestMonitoringSoftware/requirements.txt
numpy
Pillow
opencv-python
openpyxl
ExifRead
pandas
pyinstaller